fix: added validation for api token

+++ frontend/src/app/services/reservation.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { throwError, Observable } from 'rxjs';
 import { environment } from '../../environments/environments';
 import { AuthService } from './auth.service';
 
@@ -12,15 +12,28 @@ export class ReservationService {
 
   constructor(private http: HttpClient, private authService: AuthService) { }
 
+  private checkAuthOrError(): boolean {
+    return this.authService.isLoggedIn();
+  }
+
   createReservation(reservation: any): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.post(this.apiUrl, reservation, { headers: this.authService.getAuthHeaders() });
   }
 
   updateReservation(reservation: any, reservationId: number): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.patch(this.apiUrl + '/' + reservationId, reservation, { headers: this.authService.getAuthHeaders() });
   }
 
   getReservationByStructureId(id: number): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.get(
       `${this.apiUrl}/structure/${id}`,
       { headers: this.authService.getAuthHeaders() }
@@ -28,20 +41,32 @@ export class ReservationService {
   }
 
   getMonthlyReservation(id_structure: number): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.get(`${this.apiUrl}/monthly/${id_structure}`, { headers: this.authService.getAuthHeaders() });
   }
 
   getReservationById(id_structure: number): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.get(`${this.apiUrl}/${id_structure}`, { headers: this.authService.getAuthHeaders() });
   }
 
   updateReservationStatus(reservationId: number, status: any): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     const url = `${this.apiUrl}/${reservationId}/status`;
     const body = { status };
     return this.http.put(url, body, { headers: this.authService.getAuthHeaders() });
   }
 
-  deleteReservation(reservationId: number) {
+  deleteReservation(reservationId: number): Observable<any> {
+    if (!this.checkAuthOrError()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
     return this.http.delete(this.apiUrl + '/' + reservationId, { headers: this.authService.getAuthHeaders() });
   }
 }
